extract common runner logic
